Created sudoku_solver.cpp
This program solves a Sudoku puzzle by filling the empty cells.

We use backtracking to solve this problem. The basic idea is to first fill an empty cell with a number from 1 to 9. After filling it, we check if the board is still valid. If yes, we continue on to the next cell. If at any step, we get an invalid board, we move back to the previous cell and check with another number and so on.

As it may seem obvious, this algorithm is exponential with base 9, to the power of how many empty cells are there.

Therefore, T(n) = O(9^n), n is number of empty cells.